MUX PCB and schematic files
This is the MUX PCB and schematic files for KICAD. The MUX PCB was changed to include load resistors which did not get incorporated. You might not find the PCB to be as helpful to look at as you might want. This is because I skipped properly doing all the net classes and making a proper schematic.